DLCs_Main: Split the AT&T code out of DLCs.cpp.
This requierd some refactoring:
- Common variables are no longer static and are declared in DLCs.h.
- ATT.cpp has the various AT&T functions. Only those functions that
  are actually called by DLCs.cpp are in the ATT.h header; the others
  are now marked static.
- Marked one variable in DLC_ATT.h as extern so we can use it in
  DLCs.cpp without including the whole file.
