+++ DLCs_Main/ATT.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <SADXStructs.h>
+extern const PVMEntry ATT1Textures;
+extern const PVMEntry ATT2Textures;
+extern const PVMEntry ATT3Textures;
+
+void ATTTimer_Load(ObjectMaster *a1);
+void ATTObject_Load(ObjectMaster *a1);
