+++ DLCs_Main/DLCs_Main.vcxproj.filters
@@ -36,6 +36,9 @@
     <ClCompile Include="Timer.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ATT.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Balloon.h">
@@ -78,6 +81,12 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="stdafx.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="DLCs.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ATT.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
